requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.23
pillow>=9.5
tqdm>=4.66
scikit-image>=0.22
matplotlib>=3.7
